backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
pydantic==2.4.2
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
redis==4.3.4
sentence-transformers==2.2.2
faiss-cpu==1.8.0
